+++ familyhelper-webapi-node/src/main/java/com/dwarfeng/familyhelper/webapi/node/configuration/system/BeanTransformerConfiguration.java
@@ -1,13 +1,7 @@
 package com.dwarfeng.familyhelper.webapi.node.configuration.system;
 
-import com.dwarfeng.acckeeper.sdk.bean.entity.JSFixedFastJsonLoginHistory;
-import com.dwarfeng.acckeeper.sdk.bean.entity.JSFixedFastJsonLoginParamRecord;
-import com.dwarfeng.acckeeper.sdk.bean.entity.JSFixedFastJsonLoginState;
-import com.dwarfeng.acckeeper.sdk.bean.entity.JSFixedFastJsonProtectDetailRecord;
-import com.dwarfeng.acckeeper.stack.bean.entity.LoginHistory;
-import com.dwarfeng.acckeeper.stack.bean.entity.LoginParamRecord;
-import com.dwarfeng.acckeeper.stack.bean.entity.LoginState;
-import com.dwarfeng.acckeeper.stack.bean.entity.ProtectDetailRecord;
+import com.dwarfeng.acckeeper.sdk.bean.entity.*;
+import com.dwarfeng.acckeeper.stack.bean.entity.*;
 import com.dwarfeng.familyhelper.webapi.sdk.bean.disp.system.FastJsonDispAccount;
 import com.dwarfeng.familyhelper.webapi.sdk.bean.disp.system.FastJsonDispPermissionGroup;
 import com.dwarfeng.familyhelper.webapi.sdk.bean.vo.system.FastJsonAccount;
@@ -91,5 +85,10 @@ public class BeanTransformerConfiguration {
     public BeanTransformer<ProtectDetailRecord, JSFixedFastJsonProtectDetailRecord>
     protectDetailRecordBeanTransformer() {
         return new DozerBeanTransformer<>(ProtectDetailRecord.class, JSFixedFastJsonProtectDetailRecord.class, mapper);
+    }
+
+    @Bean("system.deriveHistoryBeanTransformer")
+    public BeanTransformer<DeriveHistory, JSFixedFastJsonDeriveHistory> deriveHistoryBeanTransformer() {
+        return new DozerBeanTransformer<>(DeriveHistory.class, JSFixedFastJsonDeriveHistory.class, mapper);
     }
 }